Refactor error boundary
