+++ packages/client-core/src/common/components/ErrorBoundary.tsx
@@ -25,41 +25,9 @@ Ethereal Engine. All Rights Reserved.
 
 import React from 'react'
 
-type Props = {
-  children: React.ReactNode
-}
+import { createErrorBoundary } from '@etherealengine/common/src/utils/createErrorBoundary'
 
-type ErrorHandler = (error: Error, info: React.ErrorInfo) => void
-type ErrorHandlingComponent<Props> = (props: Props, error?: Error) => React.ReactNode
-
-type ErrorState = { error?: Error }
-
-function Catch<Props extends object>(
-  component: ErrorHandlingComponent<Props>,
-  errorHandler?: ErrorHandler
-): React.ComponentType<Props> {
-  return class extends React.Component<Props, ErrorState> {
-    state: ErrorState = {
-      error: undefined
-    }
-
-    static getDerivedStateFromError(error: Error) {
-      return { error }
-    }
-
-    componentDidCatch(error: Error, info: React.ErrorInfo) {
-      if (errorHandler) {
-        errorHandler(error, info)
-      }
-    }
-
-    render() {
-      return component(this.props, this.state.error)
-    }
-  }
-}
-
-const ErrorBoundary = Catch(function error(props: Props, error?: Error) {
+const ErrorBoundary = createErrorBoundary(function error(props, error?: Error) {
   if (error) {
     return (
       <div className="error-screen">
